+++ xar/xar/lib/b64.h
@@ -7,7 +7,7 @@
 #ifndef _XAR_BASE64_H_
 #define _XAR_BASE64_H_
 
-unsigned char* xar_to_base64(const unsigned char* input, int len);
-unsigned char* xar_from_base64(const unsigned char* input, int inputLength, unsigned int *outputLength);
+unsigned char* xar_to_base64(const unsigned char* input, size_t len);
+unsigned char* xar_from_base64(const unsigned char* input, size_t inputLength, size_t *outputLength);
 
 #endif /* _XAR_BASE64_H_ */
